refactor: 다른 멘티분이 작업한 미완성 코드 레거시로 이동 및 삭제
