--- common/src/main/java/com/flab/common/auth/authservice/AuthenticationService.java
@@ -1,11 +1,0 @@
-package com.flab.common.auth.authservice;
-
-import com.flab.common.auth.AuthenticatedMember;
-
-public interface AuthenticationService {
-
-    String login(AuthenticatedMember loginMemberInfo);
-
-    void logout();
-
-}
