--- http-test/seller-api.http
@@ -1,22 +1,0 @@
-### 판매자 회원가입
-POST http://localhost:8080/api/v1/seller/signup
-Content-Type: application/json
-
-{
-  "name": "판매자1",
-  "businessNo": "1234123456",
-  "email": "audrn@gmail.com",
-  "password": "q1w2e3r4@"
-}
-
-### 판매자 로그인
-POST http://localhost:8080/api/v1/seller/login
-Content-Type: application/json
-
-{
-  "email": "audrn@gmail.com",
-  "password":"q1w2e3r4@"
-}
-
-### 가게 찾기
-GET http://localhost:8080/api/v1/seller/1
