+++ legacy/src/main/java/com/flab/livecommerce/seller/application/command/CreateSellerCommand.java
