+++ legacy/src/main/java/com/flab/livecommerce/seller/application/command/LoginSellerCommand.java
