+++ legacy/src/main/java/com/flab/livecommerce/seller/domain/Seller.java
@@ -1,15 +1,13 @@
-package com.flab.livecommerce.seller.domain;
+package com.flab.seller.domain;
 
+import com.flab.common.auth.Role;
 import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
 public class Seller {
 
@@ -19,15 +17,55 @@ public class Seller {
     private String name;
     private String businessNo;
     private String email;
+    private String password;
 
-    public Seller(String name, String businessNo, String email) {
+    @Enumerated(EnumType.STRING)
+    private Role role;
+
+    protected Seller() {
+    }
+
+    public Seller(String name, String businessNo, String email, String password) {
         this.name = name;
         this.businessNo = businessNo;
         this.email = email;
+        this.password = password;
+        this.role = Role.SELLER;
+    }
+
+    public Long getId() {
+        return id;
     }
 
     public Seller setId(Long id) {
         this.id = id;
         return this;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getBusinessNo() {
+        return businessNo;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public Role getRole() {
+        return role;
+    }
+
+    public AuthenticatedMember toLoginInfo() {
+        return new AuthenticatedMember.Builder(id)
+                .setEmail(email)
+                .setRole(role)
+                .build();
     }
 }
