+++ legacy/src/main/java/com/flab/livecommerce/seller/domain/exception/DuplicatedSellerEmailException.java
@@ -1,9 +1,9 @@
 package com.flab.seller.domain.exception;
 
-import com.flab.common.exception.BaseException;
+import com.flab.common.exception.SystemException;
 import com.flab.common.exception.ErrorCode;
 
-public class DuplicatedSellerEmailException extends BaseException {
+public class DuplicatedSellerEmailException extends SystemException {
 
     public DuplicatedSellerEmailException() {
         super(ErrorCode.DUPLICATED_EMAIL);
