+++ legacy/src/main/java/com/flab/livecommerce/seller/domain/exception/InvalidSellerException.java
@@ -1,9 +1,9 @@
 package com.flab.seller.domain.exception;
 
-import com.flab.common.exception.BaseException;
+import com.flab.common.exception.SystemException;
 import com.flab.common.exception.ErrorCode;
 
-public class InvalidSellerException extends BaseException {
+public class InvalidSellerException extends SystemException {
 
     public InvalidSellerException() {
         super(ErrorCode.INVALID_SELLER);
