+++ legacy/src/main/java/com/flab/livecommerce/seller/infrastructure/config/SellerRepositoryConfig.java
