+++ legacy/src/main/java/com/flab/livecommerce/seller/infrastructure/encryption/SellerSecurityPasswordEncoder.java
