+++ legacy/src/main/java/com/flab/livecommerce/seller/infrastructure/jpaconfig/SellerJpaConfig.java
