+++ legacy/src/main/java/com/flab/livecommerce/seller/infrastructure/sessionproperties/SessionProperties.java
