+++ legacy/src/main/java/com/flab/livecommerce/seller/presentation/request/CreateSellerRequest.java
