+++ legacy/src/main/java/com/flab/livecommerce/seller/presentation/response/LoginSellerResponse.java
