--- seller-api/seller-application/build.gradle
@@ -1,15 +1,0 @@
-plugins {
-    id 'io.spring.dependency-management' version '1.0.12.RELEASE'
-    id 'org.springframework.boot' version '2.7.2'
-}
-
-dependencies {
-    implementation project(path: ':seller-api:seller-domain', configuration: "default")
-    implementation "org.springframework.boot:spring-boot-starter-data-jpa"
-    compileOnly 'org.apache.tomcat:tomcat-servlet-api:9.0.37'
-    testImplementation "org.springframework.boot:spring-boot-starter-test"
-}
-
-tasks.getByName('bootJar') {
-    enabled = false
-}
