--- seller-api/seller-application/src/main/java/com/flab/seller/application/SearchSellerProcessor.java
@@ -1,17 +1,0 @@
-package com.flab.seller.application;
-
-import com.flab.seller.domain.Seller;
-import com.flab.seller.domain.SellerRepository;
-
-public class SearchSellerProcessor {
-
-    private final SellerRepository sellerRepository;
-
-    public SearchSellerProcessor(SellerRepository sellerRepository) {
-        this.sellerRepository = sellerRepository;
-    }
-
-    public Seller execute(Long id) {
-        return sellerRepository.findById(id);
-    }
-}
