--- seller-api/seller-domain/src/main/java/com/flab/seller/domain/SellerRepository.java
@@ -1,18 +1,0 @@
-package com.flab.seller.domain;
-
-
-
-public interface SellerRepository {
-
-    Seller save(Seller seller);
-
-    Seller findById(Long id);
-
-    Seller findByEmail(String email);
-
-    boolean existsByEmail(String email);
-
-    boolean existsByName(String name);
-
-    boolean existsByBusinessNo(String businessNo);
-}
