--- seller-api/seller-infrastructure/build.gradle
@@ -1,19 +1,0 @@
-plugins {
-    id 'io.spring.dependency-management' version '1.0.12.RELEASE'
-    id 'org.springframework.boot' version '2.7.2'
-}
-
-dependencies {
-    implementation project(path: ':seller-api:seller-domain', configuration: "default")
-    implementation project(path: ':seller-api:seller-application', configuration: "default")
-    implementation project(path: ':seller-api:seller-presentation', configuration: "default")
-    implementation "org.springframework.boot:spring-boot-starter-web"
-    implementation 'org.springframework.security:spring-security-crypto'
-    implementation 'org.springframework.boot:spring-boot-starter-data-redis'
-    implementation "org.springframework.boot:spring-boot-starter-data-jpa"
-    testImplementation "org.springframework.boot:spring-boot-starter-test"
-}
-
-tasks.getByName('bootJar') {
-    enabled = false
-}
